+++ backend/services/task.go
@@ -18,11 +18,11 @@ import (
 	"github.com/spf13/viper"
 	"os"
 	"os/exec"
+	"path"
 	"path/filepath"
 	"runtime"
 	"runtime/debug"
 	"strconv"
-	"strings"
 	"sync"
 	"syscall"
 	"time"
@@ -110,14 +110,11 @@ func AssignTask(task model.Task) error {
 func SetEnv(cmd *exec.Cmd, envs []model.Env, taskId string, dataCol string) *exec.Cmd {
 	// 默认把Node.js的全局node_modules加入环境变量
 	envPath := os.Getenv("PATH")
-	for _, _path := range strings.Split(envPath, ":") {
-		if strings.Contains(_path, "/.nvm/versions/node/") {
-			pathNodeModules := strings.Replace(_path, "/bin", "/lib/node_modules", -1)
-			_ = os.Setenv("PATH", pathNodeModules+":"+envPath)
-			_ = os.Setenv("NODE_PATH", pathNodeModules)
-			break
-		}
-	}
+	homePath := os.Getenv("HOME")
+	nodeVersion := "v8.12.0"
+	nodePath := path.Join(homePath, ".nvm/versions/node", nodeVersion, "lib/node_modules")
+	_ = os.Setenv("PATH", nodePath+":"+envPath)
+	_ = os.Setenv("NODE_PATH", nodePath)
 
 	// 默认环境变量
 	cmd.Env = append(os.Environ(), "CRAWLAB_TASK_ID="+taskId)
@@ -246,7 +243,6 @@ func ExecuteShellCmd(cmdStr string, cwd string, t model.Task, s model.Spider) (e
 	if runtime.GOOS == constants.Windows {
 		cmd = exec.Command("cmd", "/C", cmdStr)
 	} else {
-		cmd = exec.Command("")
 		cmd = exec.Command("sh", "-c", cmdStr)
 	}
 
@@ -354,10 +350,9 @@ func SaveTaskResultCount(id string) func() {
 func ExecuteTask(id int) {
 	if flag, ok := LockList.Load(id); ok {
 		if flag.(bool) {
-			log.Debugf(GetWorkerPrefix(id) + "正在执行任务...")
+			log.Debugf(GetWorkerPrefix(id) + "running tasks...")
 			return
 		}
-
 	}
 
 	// 上锁
@@ -382,6 +377,7 @@ func ExecuteTask(id int) {
 
 	// 节点队列
 	queueCur := "tasks:node:" + node.Id.Hex()
+
 	// 节点队列任务
 	var msg string
 	if msg, err = database.RedisClient.LPop(queueCur); err != nil {
@@ -391,6 +387,7 @@ func ExecuteTask(id int) {
 		}
 	}
 
+	// 如果没有获取到任务，返回
 	if msg == "" {
 		return
 	}
@@ -508,6 +505,8 @@ func ExecuteTask(id int) {
 		log.Errorf(GetWorkerPrefix(id) + err.Error())
 		return
 	}
+
+	// 统计数据
 	t.Status = constants.StatusFinished                     // 任务状态: 已完成
 	t.FinishTs = time.Now()                                 // 结束时间
 	t.RuntimeDuration = t.FinishTs.Sub(t.StartTs).Seconds() // 运行时长
@@ -669,7 +668,7 @@ func CancelTask(id string) (err error) {
 	return nil
 }
 
-func AddTask(t model.Task) error {
+func AddTask(t model.Task) (string, error) {
 	// 生成任务ID
 	id := uuid.NewV4()
 	t.Id = id.String()
@@ -686,17 +685,17 @@ func AddTask(t model.Task) error {
 	if err := model.AddTask(t); err != nil {
 		log.Errorf(err.Error())
 		debug.PrintStack()
-		return err
+		return t.Id, err
 	}
 
 	// 加入任务队列
 	if err := AssignTask(t); err != nil {
 		log.Errorf(err.Error())
 		debug.PrintStack()
-		return err
+		return t.Id, err
 	}
 
-	return nil
+	return t.Id, nil
 }
 
 func GetTaskEmailMarkdownContent(t model.Task, s model.Spider) string {
@@ -850,6 +849,14 @@ func SendNotifications(u model.User, t model.Task, s model.Spider) {
 		go func() {
 			SendTaskWechat(u, t, s)
 		}()
+	}
+}
+
+func UnlockLongTask(s model.Spider, n model.Node) {
+	if s.IsLongTask {
+		colName := "long-tasks"
+		key := fmt.Sprintf("%s:%s", s.Id.Hex(), n.Id.Hex())
+		_ = database.RedisClient.HDel(colName, key)
 	}
 }