+++ frontend/src/utils/index.js
@@ -1,9 +1,13 @@
 import stats from './stats'
 import encrypt from './encrypt'
 import tour from './tour'
+import log from './log'
+import scrapy from './scrapy'
 
 export default {
   stats,
   encrypt,
-  tour
+  tour,
+  log,
+  scrapy
 }